requirements.txt
scipy==1.5.3
numpy==1.18.4
pandas==1.0.4
matplotlib==3.2.1
nltk==3.5
Pillow==8.0.1
plotly==4.12.0
seaborn==0.11.0
streamlit==0.71.0
wordcloud==1.8.1
